backend/requirements-test.txt
# Testing Dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-asyncio==0.21.1
pytest-flask==1.3.0
coverage==7.3.2

# Code Quality
flake8==6.1.0
black==23.11.0
isort==5.12.0
mypy==1.7.1

# Security
bandit==1.7.5
safety==2.3.5



